+++ PatchDay/Hormones/HormonesViewController.swift
@@ -64,7 +64,12 @@ class HormonesViewController: UIViewController, UITableViewDataSource, UITableVi
 	}
 
 	func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-		viewModel.handleRowTapped(at: indexPath.row, self)
+		viewModel.handleRowTapped(at: indexPath.row, self) {
+			DispatchQueue.main.async {
+				self.initViewModel()
+				self.hormonesTableView.reloadRows(at: [indexPath], with: .none)
+			}
+		}
 	}
 
 	// MARK: - Actions